Switch producto precios to float and add migration
Convert producto price columns and related code from Decimal/Numeric to float/Float. Changes in app/db/models.py update precio_1/2/3 and precio_venta types; app/services/inventario.py adjusts establecer_precio signature and casts to float (margin still computed via Decimal(str(precio_1)) to preserve logic); app/ui/producto_form_dialog.py returns floats for price getters. Added migrations/0048_agregar_precio_1_2_3.sql to rename/add precio_1/2/3 columns, populate defaults and set NOT NULL. Run the migration and test for potential precision differences.

+++ app/ui/producto_form_dialog.py
@@ -562,10 +562,10 @@ class ProductoFormDialog(QDialog):
             ),
         }
 
-    def get_precio_venta(self) -> Decimal:
+    def get_precio_venta(self) -> float:
         return self.precio_1_input.get_value()
 
-    def get_precios(self) -> tuple[Decimal, Decimal, Decimal]:
+    def get_precios(self) -> tuple[float, float, float]:
         return (
             self.precio_1_input.get_value(),
             self.precio_2_input.get_value(),
